# server/docs/swagger.yaml
basePath: /
definitions:
...
  flow_schema.FlowApplyResp:
    properties:
      applyUserId:
        description: 申请人id
        type: string
      applyUserNickname:
        description: 申请人昵称
        type: string
      createTime:
        description: 创建时间
        type: string
      flowFormData:
        description: 表单配置
        type: string
      flowGroup:
        description: 流程分类
        type: integer
      flowName:
        description: 流程名称
        type: string
      flowProcessData:
        description: 流程配置
        type: string
      flowProcessDataList:
        description: 流程配置list数据
        type: string
      flowRemark:
        description: 流程描述
        type: string
      formValue:
        description: 表单值
        type: string
      id:
        type: string
      status:
        description: 状态：1待提交，2审批中，3审批完成，4审批失败
        type: integer
      templateId:
        description: 模板
        type: string
      updateTime:
        description: 更新时间
        type: string
    type: object
...
  response.PageResp:
    properties:
      count:
        description: 总数
        type: integer
      lists:
        description: 数据列表
      pageNo:
        description: 当前页码
        type: integer
      pageSize:
        description: 每页数量
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact:
    email: x@adtk.cn
    name: xh
    url: http://x.adtk.cn
  description: x_admin后台管理系统
  license:
    name: MIT License
    url: https://gitee.com/xiangheng/x_admin/blob/main/LICENSE
  termsOfService: http://x.adtk.cn
paths:
...
  /api/admin/flow/flow_apply/add:
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 模板
        in: body
        name: templateId
        schema:
          type: string
      - description: 申请人id
        in: body
        name: applyUserId
        schema:
          type: string
      - description: 申请人昵称
        in: body
        name: applyUserNickname
        schema:
          type: string
      - description: 流程名称
        in: body
        name: flowName
        schema:
          type: string
      - description: 表单值
        in: body
        name: formValue
        schema:
          type: string
      - description: 状态：1待提交，2审批中，3审批完成，4审批失败
        in: body
        name: status
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 申请流程新增
      tags:
      - flow_apply-申请流程
  /api/admin/flow/flow_apply/del:
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 申请id
        in: body
        name: id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 申请流程删除
      tags:
      - flow_apply-申请流程
  /api/admin/flow/flow_apply/detail:
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 申请id
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/flow_schema.FlowApplyResp'
              type: object
      summary: 申请流程详情
      tags:
      - flow_apply-申请流程
  /api/admin/flow/flow_apply/edit:
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 申请id
        in: body
        name: id
        schema:
          type: string
      - description: 流程名称
        in: body
        name: flowName
        schema:
          type: string
      - description: 表单值
        in: body
        name: formValue
        schema:
          type: string
      - description: 状态：1待提交，2审批中，3审批完成，4审批失败
        in: body
        name: status
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 申请流程编辑
      tags:
      - flow_apply-申请流程
  /api/admin/flow/flow_apply/list:
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 页码
        in: query
        name: pageNo
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 模板
        in: query
        name: templateId
        type: string
      - description: 申请人id
        in: query
        name: applyUserId
        type: string
      - description: 申请人昵称
        in: query
        name: applyUserNickname
        type: string
      - description: 流程名称
        in: query
        name: flowName
        type: string
      - description: 流程分类
        in: query
        name: flowGroup
        type: integer
      - description: 流程描述
        in: query
        name: flowRemark
        type: string
      - description: 表单配置
        in: query
        name: flowFormData
        type: string
      - description: 流程配置
        in: query
        name: flowProcessData
        type: string
      - description: 流程配置list数据
        in: query
        name: flowProcessDataList
        type: string
      - description: 表单值
        in: query
        name: formValue
        type: string
      - description: 状态：1待提交，2审批中，3审批完成，4审批失败
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResp'
                  - properties:
                      lists:
                        items:
                          $ref: '#/definitions/flow_schema.FlowApplyResp'
                        type: array
                    type: object
              type: object
      summary: 申请流程列表
      tags:
      - flow_apply-申请流程
...
swagger: "2.0"